+++ projects/backbone/app/controllers/application_controller.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class ApplicationController < ActionController::Base
+  protect_from_forgery with: :exception
+
+  def home
+    head(:ok)
+  end
+end