+++ projects/backbone/app/javascript/utils/index.test.ts
@@ -1,0 +1,5 @@
+describe("foo", () => {
+  it("bar", () => {
+    expect(2).toBe(2);
+  });
+});