+++ projects/backbone/config/routes.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+Rails.application.routes.draw do
+  resources :command_events
+
+  # Web app
+  root to: "application#home"
+end