+++ projects/backbone/app/javascript/entrypoints/application.js
@@ -1,0 +1,7 @@
+require("@rails/ujs").start();
+
+// Styles
+require("../styles/application.css");
+
+require("chartkick")
+require("chart.js")