+++ projects/cloud/app/controllers/users/sessions_controller.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Users
+  class SessionsController < Devise::SessionsController
+    protected
+      def respond_to_on_destroy
+        head(:no_content)
+      end
+  end
+end