+++ projects/cloud/app/frontend/utilities/signOut.ts
@@ -1,0 +1,9 @@
+export const redirectToSignIn = () => {
+  window.location.pathname = '/users/sign_in';
+};
+
+const signOut = async () => {
+  redirectToSignIn();
+};
+
+export default signOut;