+++ projects/cloud/app/graphql/types/user_type.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Types
+  class UserType < Types::BaseObject
+    field :id, ID, null: false
+    field :email, String, null: false
+    field :avatar_url, String, null: true
+    field :projects, [ProjectType], null: false
+  end
+end